Share NDJSON distinct keys
